Add scripts and .meta files
Due to Github size limits, it is inappropriate to upload the entire
project. This would also be a potential copyright violation in the
case of paid assets, such as the car.

The .meta files are used to assign each asset an ID, so that it can
be referred to in the editor. If the .meta files are lost, Unity
would reimport each asset with a new ID, breaking all references.

Therefore, theoretically, if I just commit the meta files but not
the large assets, it would be possible to reconstruct the project
by placing each asset to match its .meta file. I haven't tried this,
but it should be possible!